backend/requirements.txt
Django>=4.2,<5.1
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
djangorestframework>=3.14
django-filter>=23.2
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.26
django-environ>=0.11
django-cors-headers>=4.2
Pillow>=10.0
python-dateutil>=2.8
